Refactor: Move all editor (socket) related code to its own module
