+++ daemon/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod actors;
 pub mod connect;
 pub mod daemon;
+pub mod editor;
 pub mod logging;
 pub mod ot;
 pub mod peer;
